Support importing existing Kafka clusters (fixes #64) (#74)
This change should address the issue reported in #64 wherein it's not currently possible to import existing Kafka clusters.

The issue is that the read function for clusters requires both the cluster ID and the containing environment ID, which means that using the pass-through state importer function doesn't work.

I've addressed this by allowing clusters to be imported by passing an import ID of the form `<environment ID>/<cluster ID>`.